qgs3dsceneexporter: Add support for stl export format
The existing `Qgs3DSceneExporter::save` function has a new parameter
`exportFormat` to choose between `Stl` and `Obj`. It is split in two
functions `saveObj` which saves in Obj format and reuses the
existing code from `save` and the new `saveStl` function which saves
an object in stl format.

The new `saveStl` is similar to the existing `saveObj` one. One
iterate over the different objects of the scene and call
`Qgs3dExportObject::saveTo`. The `stl` logic is simpler to the the
`obj` one because, `stl` cannot save textures.