qgs3dexportobject: Add stl export format
The existing `Qgs3DExportObject::saveTo` function has a new parameter
`exportFormat` to choose between `Stl` and `Obj`. It is split in two
functions `saveToObj` which saves in Obj format and reuses the
existing code from `saveTo` and the new `saveToStl` function which
saves an object in stl format. The `stl` format is simpler than the
`obj` one. One can not save a texture and one cannot smooth edges and
save normals.